Finished by typed all entities with 'DTO'.

+++ src/main/java/com/list/server/domain/entities/Item.java
@@ -22,7 +22,7 @@ public class Item {
     private Long id;
     @Column(nullable = false)
     private String name;
-    private int quantity;
+    private short quantity;
 
     @ManyToOne
     @JoinColumn(name = "category_id")
